YAWLEditor 1.4.4 tag release added (exactly as per the Zip file that Moe shipped to Sourceforge).
This is not a typical 'diff' branch of a subversion revision. Moe built this image from the 1.4.3 source on Sourceforge, adding her own stuff and then committing the new build to Sourceforge, leaving the subversion repository out of the loop (there was a clash in timing on IHG compatability issues and Moe's need to release a version for a paper).